fixed most warnings & refactoring

+++ src/components/Vision.hpp
@@ -1,11 +1,11 @@
 #pragma once
 
-#include"../constants.hpp"
+#include "../constants.hpp"
 #include <vector>
 
 struct Vision {
-	int range = 5 * TILE_SIZE;						// in pixels
-	int angle = 180;								// FOV in degrees
+	float range = 5 * TILE_SIZE; // in pixels
+	float angle = 180;           // FOV in degrees
 	std::vector<Entity> visibleEnemies = {};
 	std::vector<Entity> visibleAllies = {};
